Server: refactor rss.py and watchlist.py
